requirements.txt
# JetBot API Requirements
# Based on JetBot Jupyter container dependencies

# Core dependencies
websockets
# opencv-python  # Use system opencv instead (has GStreamer support)
# NOTE: pip opencv-python does NOT have GStreamer/nvargus support needed for Jetson camera
numpy<2  # Use latest compatible version (system OpenCV handles compatibility)
traitlets
ipywidgets
Adafruit_MotorHAT
smbus2
pydantic
fastapi

# JetBot dependencies (from container)
sparkfun_qwiic==2.0.0
sparkfun_qwiic_ads1015==2.0.0
sparkfun_qwiic_alphanumeric==2.0.0
sparkfun_qwiic_as7265x==2.0.0
sparkfun_qwiic_as726x==2.0.0
sparkfun_qwiic_bme280==2.0.0
sparkfun_qwiic_bmp581==2.0.0
sparkfun_qwiic_button==2.0.3
sparkfun_qwiic_cap1203==2.0.0
sparkfun_qwiic_ens160==2.0.0
sparkfun_qwiic_fs3000==2.0.0
sparkfun_qwiic_gpio==2.0.0
sparkfun_qwiic_huskylens==2.0.0
sparkfun_qwiic_i2c==2.0.0
sparkfun_qwiic_icm20948==2.0.1
sparkfun_qwiic_ism330dhcx==2.0.0
sparkfun_qwiic_keypad==2.0.0
sparkfun_qwiic_kx13x==2.0.0
sparkfun_qwiic_led_stick==2.0.0
sparkfun_qwiic_lsm6dso==2.0.0
sparkfun_qwiic_max1704x==2.0.0
sparkfun_qwiic_max3010x==2.0.0
sparkfun_qwiic_mcp4725==2.0.0
sparkfun_qwiic_mcp9600==2.0.0
sparkfun_qwiic_micropressure==2.0.0
sparkfun_qwiic_mmc5983ma==2.0.0
sparkfun_qwiic_nau7802==2.0.0
sparkfun_qwiic_oled_base==2.0.0
sparkfun_qwiic_otos==2.0.0
sparkfun_qwiic_pca9685==2.0.0
sparkfun_qwiic_rfid==2.1.1
sparkfun_qwiic_rv8803==2.0.0
sparkfun_qwiic_scd4x==2.0.0
sparkfun_qwiic_scmd==2.0.0
sparkfun_qwiic_serlcd==2.0.0
sparkfun_qwiic_sgp40==2.0.0
sparkfun_qwiic_tca9548a==2.0.0
sparkfun_qwiic_titan_gps==2.0.1
sparkfun_qwiic_tmp102==2.0.0
sparkfun_qwiic_tmp117==2.0.1
sparkfun_qwiic_ultrasonic==2.0.0
sparkfun_qwiic_veml6030==2.0.0
sparkfun_qwiic_vl53l1x==2.0.0
sparkfun_qwiic_vl53l5cx==2.0.0

# Additional Qwiic packages
sparkfun-qwiic-adxl313==0.0.7
sparkfun-qwiic-as6212==0.0.2
sparkfun-qwiic-ccs811==0.9.4
sparkfun-qwiic-dual-encoder-reader==0.0.2
sparkfun-qwiic-eeprom==0.0.1
sparkfun-qwiic-joystick==2.0.0
sparkfun-qwiic-pir==0.0.4
sparkfun-qwiic-proximity==0.9.0
sparkfun-qwiic-relay==0.0.2
sparkfun-qwiic-soil-moisture-sensor==0.0.2
sparkfun-qwiic-twist==2.0.0

# Adafruit CircuitPython packages
Adafruit-Blinka==8.66.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-motor==3.4.17
adafruit-circuitpython-motorkit==1.6.18
adafruit-circuitpython-pca9685==3.4.19
adafruit-circuitpython-register==1.10.4
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.83.1
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2

# Additional dependencies
pyserial==3.5
spidev==3.8
sysv-ipc==1.1.0
pyftdi==0.57.1
pyusb==1.3.1
binho-host-adapter==0.1.6

# PyTorch (install separately with Jetson-optimized version)
# pip install torch==2.8.0 torchvision==0.23.0 torchaudio==2.8.0 --index-url https://download.pytorch.org/whl/cu124

# Additional image/scientific packages
matplotlib==3.10.6
matplotlib-inline==0.1.7
Pillow==9.0.1

# System dependencies that need to be installed via apt:
# sudo apt install -y python3-gi python3-gi-cairo gir1.2-gstreamer-1.0 python3-gst-1.0
# sudo apt install -y libcairo2-dev libgirepository1.0-dev pkg-config libcairo-gobject2
# sudo apt install -y gstreamer1.0-plugins-base gstreamer1.0-plugins-good gstreamer1.0-plugins-bad
# sudo apt install -y libopencv-dev python3-opencv
# sudo apt install -y python3-libnvinfer python3-libnvinfer-dev python3-libnvinfer-dispatch python3-libnvinfer-lean

# Note: PyGObject is installed via system packages, not pip
# Note: TensorRT packages are NVIDIA-specific and installed via apt (not pip)
# Note: Some packages may conflict - use --no-deps flag for problematic packages if needed

# ADK Backend
fastapi
google-adk
litellm
google-genai

# YOLO-E Backend
# Note: ultralytics installed via setup_dependencies.sh with --no-deps to avoid opencv conflict
# Note: OpenCV and PyTorch installed via setup_dependencies.sh
# Ultralytics dependencies (excluding opencv and torch which are handled separately)
PyYAML
requests
matplotlib
seaborn
tqdm
psutil
pandas